+++ spec/rubocop/options_spec.rb
@@ -52,6 +52,9 @@ RSpec.describe RuboCop::Options, :isolated_environment do
                                                files are present in the directory tree.
                   --auto-gen-config            Generate a configuration file acting as a
                                                TODO list.
+                  --regenerate-todo            Regenerate the TODO configuration file using
+                                               the last configuration. If there is no existing
+                                               TODO file, acts like --auto-gen-config.
                   --exclude-limit COUNT        Used together with --auto-gen-config to
                                                set the limit for how many Exclude
                                                properties to generate. Default is 15.
@@ -367,6 +370,75 @@ RSpec.describe RuboCop::Options, :isolated_environment do
       it 'accepts other options' do
         expect { options.parse %w[--auto-gen-config --lint] }
           .not_to raise_error
+      end
+    end
+
+    describe '--regenerate-todo' do
+      subject(:parsed_options) { options.parse(command_line_options).first }
+
+      let(:config_regeneration) do
+        instance_double(RuboCop::ConfigRegeneration, options: todo_options)
+      end
+      let(:todo_options) { { auto_gen_config: true, exclude_limit: '100', offense_counts: false } }
+
+      before do
+        allow(RuboCop::ConfigRegeneration).to receive(:new).and_return(config_regeneration)
+      end
+
+      context 'when no other options are given' do
+        let(:command_line_options) { %w[--regenerate-todo] }
+        let(:expected_options) do
+          {
+            auto_gen_config: true,
+            exclude_limit: '100',
+            offense_counts: false,
+            regenerate_todo: true
+          }
+        end
+
+        it { is_expected.to eq(expected_options) }
+      end
+
+      context 'when todo options are overridden before --regenerate-todo' do
+        let(:command_line_options) { %w[--exclude-limit 50 --regenerate-todo] }
+        let(:expected_options) do
+          {
+            auto_gen_config: true,
+            exclude_limit: '50',
+            offense_counts: false,
+            regenerate_todo: true
+          }
+        end
+
+        it { is_expected.to eq(expected_options) }
+      end
+
+      context 'when todo options are overridden after --regenerate-todo' do
+        let(:command_line_options) { %w[--regenerate-todo --exclude-limit 50] }
+        let(:expected_options) do
+          {
+            auto_gen_config: true,
+            exclude_limit: '50',
+            offense_counts: false,
+            regenerate_todo: true
+          }
+        end
+
+        it { is_expected.to eq(expected_options) }
+      end
+
+      context 'when disabled options are overridden to be enabled' do
+        let(:command_line_options) { %w[--regenerate-todo --offense-counts] }
+        let(:expected_options) do
+          {
+            auto_gen_config: true,
+            exclude_limit: '100',
+            offense_counts: true,
+            regenerate_todo: true
+          }
+        end
+
+        it { is_expected.to eq(expected_options) }
       end
     end