+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## master (unreleased)
 
+### New features
+
+* [#8778](https://github.com/rubocop-hq/rubocop/pull/8778): Add command line option `--regenerate-todo`. ([@dvandersluis][])
+
 ### Bug fixes
 
 * [#8774](https://github.com/rubocop-hq/rubocop/issues/8774): Fix a false positive for `Layout/ArrayAlignment` with parallel assignment. ([@dvandersluis][])
+
 
 ## 0.91.1 (2020-09-23)